Improved Java code to use a single constants file an made Util functions static
