+++ backend/src/main/java/com/example/backend/BackendApplication.java
@@ -30,53 +30,4 @@ public class BackendApplication {
     // Released under MIT License
 
 
-    @GetMapping("/players")
-    public ResponseEntity<String> getPlayers(@RequestParam(name = "startId", defaultValue = "0") Integer startId,
-                             @RequestParam(name = "endId", defaultValue = "50") Integer endId) throws ClassNotFoundException, InstantiationException, IllegalAccessException {
-        try {
-            Connection connection = DriverManager.getConnection(url, username, password);
-            String readMessageQuery = "SELECT player_id, name, birthday, height, weight FROM Players WHERE player_id >="
-                    + startId + " AND player_id < " + endId;
-            Statement statement = connection.createStatement();
-            ResultSet resultSet = statement.executeQuery(readMessageQuery);
-            JSONArray res = util.resultToJsonArray(resultSet, connection);
-
-            return ResponseEntity.ok(res.toString());
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return ResponseEntity.badRequest().body(String.format("Unable to parse JSON: %s", e));
-        }
-    }
-
-    @GetMapping("/playerCount")
-    public ResponseEntity<String> getPlayerCount() throws ClassNotFoundException, InstantiationException, IllegalAccessException {
-        try {
-            Connection connection = DriverManager.getConnection(url, username, password);
-            String readMessageQuery = "SELECT COUNT(*) AS playerCount FROM Players";
-            Statement statement = connection.createStatement();
-            ResultSet resultSet = statement.executeQuery(readMessageQuery);
-            JSONArray res = util.resultToJsonArray(resultSet, connection);
-
-            return ResponseEntity.ok(res.toString());
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return ResponseEntity.badRequest().body(String.format("Unable to parse JSON: %s", e));
-        }
-    }
-
-    @GetMapping("/player/{ID}")
-    public ResponseEntity<String> getPlayerByID(@PathVariable String ID) throws ClassNotFoundException, InstantiationException, IllegalAccessException {
-        try {
-            Connection connection = DriverManager.getConnection(url, username, password);
-            String readMessageQuery = "SELECT * FROM Players WHERE player_id = " + ID;
-            Statement statement = connection.createStatement();
-            ResultSet resultSet = statement.executeQuery(readMessageQuery);
-            JSONArray res = util.resultToJsonArray(resultSet, connection);
-            statement.close();
-            return ResponseEntity.ok(res.toString());
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return ResponseEntity.badRequest().body(String.format("Unable to parse JSON: %s", e));
-        }
-    }
 }
