+++ backend/src/main/java/com/example/backend/Constants.java
@@ -1,0 +1,8 @@
+package com.example.backend;
+
+public class Constants {
+    public static final String username = "root"; // our default username is root
+    public static final String password = "password"; // our default password is password
+    public static final String dbName = "footyfiend"; // our current db name is footy_fiend
+    public static final String url = "jdbc:mysql://127.0.0.1:3306/" + dbName;
+}
